Commit 4
Levels now work, but the code can be much improved. For example I should
run prepareForGame-function the first time as well, and not only when
you have a game over. Further, the code that runs from MasterController
can be improved also. Scores are not added between levels, and not coins
either. Score, however, works as intended.